Correct includes value in library.properties
The includes field in library.properties is used to specify which #include directives should be added to the sketch via Sketch > Include Library > ShiftRegister74HC595. This field should specify the primary header file(s) of the library. Therefore, the value should be changed from Arduino.h to ShiftRegister74HC595.h. There is no need to #include Arduino.h in the sketch because the Arduino IDE automatically does this.